% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acid.copper.pattern.plating}
\alias{acid.copper.pattern.plating}
\title{Acid copper pattern plating experiment}
\format{A data frame with 10 rows and 5 variables:
\describe{
  \item{xA}{Anode-cathode separation (in.) - uncoded}
  \item{xB}{Current Density (asf) - uncoded}
  \item{s}{standard deviation}
  \item{zA}{Anode-cathode separation (in.) - coded}
  \item{zB}{Current Density (asf) - coded}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
acid.copper.pattern.plating
}
\description{
Experiment to minimize coating thinkness variation of an acid copper-plating
process. (Chapter 16)
}
\keyword{datasets}

