% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aerosol}
\alias{aerosol}
\title{Aerosol}
\format{A data frame with 24 rows and 4 variables:
\describe{
  \item{MANUF}{Manufacturer}
  \item{FILTER}{Filter type}
  \item{AERO}{Aerosol type}
  \item{Y}{Percent penetration}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
aerosol
}
\description{
Experiment to determine whether the current standard aerosol that was
used for testing respirator filters could be replaced by an alternative
aerosol. (Chapter 18)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

