% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air.freshener}
\alias{air.freshener}
\title{Air freshener}
\format{A data frame with 32 rows and 4 variables:
\describe{
 \item{WEEK}{one-week periods}
 \item{STORE}{store}
 \item{TRTMT}{price-and-display treatments}
 \item{Y}{unit sales for a one week period}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
air.freshener
}
\description{
Experiment to compare the effects of four price-and-display treatments on the
sales of a brand of air freshener. (Chapter 12, Table 12.14)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

