% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air.rifle}
\alias{air.rifle}
\title{Air rifle}
\format{A data frame with 48 rows and 4 variables:
\describe{
 \item{BLOCK}{paper covering the target, changed after every 4 observations}
 \item{PROJ}{projectile type}
 \item{PUMPS}{number of pumps of the rifle}
 \item{Y}{distance from center of taget (mm)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
air.rifle
}
\description{
Experiment to examine the accuracy of an air rifle shooting at a target.
(Chapter 11)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

