% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air.velocity}
\alias{air.velocity}
\title{Air velocity}
\format{A data frame with 18 rows and 4 variables:
\describe{
 \item{TC}{treatment code}
 \item{HT}{height of ribs on the roughened rod}
 \item{REYNO}{Reynolds number}
 \item{VELOCITY}{\eqn{ (d - 1.4) * 10^3} where d is the distance in inches
 from center of rod}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
air.velocity
}
\description{
Experiment to examine the position of maximum velocity of air blown down
the space between a roughened rod and a smooth pipe surrounding it. (Chapter 6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

