% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol}
\format{A data frame with 24 rows and 2 variables:
\describe{
 \item{BOTTLE}{Bottle of alcohol solution}
 \item{CONC}{concentration of alcohol determined by gas chromatography (mg/ml)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
alcohol
}
\description{
Alcohol concentrations of samples of alcohol solutions taken from six bottles of
alcohol solution randomly selected from a large batch. (Chapter 17)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

