% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algorithm}
\alias{algorithm}
\title{Algorithm}
\format{A data frame with 36 rows and 6 variables:
\describe{
 \item{DAY}{day algorithm was run}
 \item{ALG}{algorithm type - 3 types}
 \item{SFRAC}{sampling fractions - 3 levels}
 \item{POP}{populations sizes - 3 levels}
 \item{Y}{execution time (in 1/00 sec)}
 \item{TC}{treatment combination}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
algorithm
}
\description{
An experiement to compare the speeds of three computer algorithms for selecting
a simple random sample of size m without replacement from a population of
size N.
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

