% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ammunition}
\alias{ammunition}
\title{Ammunition}
\format{A data frame with 32 rows and 3 variables:
\describe{
 \item{PROJTL}{projectile lot}
 \item{CHARGE}{charge lot}
 \item{VELCTY}{muzzle velocities (ft/sec)}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
ammunition
}
\description{
An experiment concerning the muzzle velocity characteristics of ammunition for a
field artillery weapon. (Chapter 17)
}
\keyword{datasets}

