% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balloon}
\alias{balloon}
\title{Balloon}
\format{A data frame with 32 rows and 3 variables:
\describe{
 \item{ORDER}{order of inflation}
 \item{COLOR}{color of balloon - four levels (pink, yellow, orange, blue)}
 \item{TIME}{time to inflate balloon (nearest 1/10 second)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
balloon
}
\description{
Experiment to determine if balloons of different colors are similar in terms
of the time taken for inflation to a diameter of 7 inches. (Chapter 3)
}
\keyword{datasets}

