% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banana}
\alias{banana}
\title{Banana}
\format{A data frame with 48 rows and 4 variables:
\describe{
 \item{LIGHT}{lighting conditions - 2 levels}
 \item{STORE}{storage method - 2 levels}
 \item{BLOCK}{the experimenter who examined the bananas - 3 levels}
 \item{Y}{percentage of blackened skin}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
banana
}
\description{
An experiment to determine if lighting conditions and storage method affect
percentage of blackened skin on bananas (Chapter 10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

