% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{battery}
\alias{battery}
\title{Battery}
\format{A data frame with 16 rows and 5 variables:
\describe{
 \item{TYPEBAT}{type of battery - four levels}
 \item{LIFE}{minutes of battery life}
 \item{UCOST}{unit cost in dollars - half of the cost of a two-pack of batteries}
 \item{LPUC}{lifetime per unit cost measured in minites per dollar}
 \item{ORDER}{time order}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
battery
}
\description{
An experiment to determine what type of nonrechargeable battery was most
economical. (Chapter 2, Table 2.8)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

