% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bean.soaking}
\alias{bean.soaking}
\title{Bean soaking}
\format{A data frame with 68 rows and 2 variables.
\describe{
 \item{TIME}{soaking times - 5 levels}
 \item{LENGTH}{length of a shoot 48 hours after soaking (mm)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
bean.soaking
}
\description{
An experiment to compare the effects of 5 different soaking times on the growth
rate of mung bean seeds. (Chapters 4 & 8, Table 8.6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

