% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beef}
\alias{beef}
\title{Beef}
\format{A data frame with 30 rows and 3 variables:
\describe{
 \item{BLOCK}{blocks}
 \item{TRTMT}{periods of storage - six levels (0,1,2,4,9, and 18 days)}
 \item{Y}{tenderness score; max score is 40}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
beef
}
\description{
An experiment to compare the effects of cold storage on the tenderness of beef
roasts. (Chapter 11, Table 11.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

