% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bicycle}
\alias{bicycle}
\title{Bicycle}
\format{A data frame with 15 rows and 2 variables:
\describe{
 \item{TRT}{speeds - 5 levels (5, 10, 15, 20 and 25 mph)}
 \item{RATE}{crank rate}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
bicycle
}
\description{
An experiment to compare crank rates required to keep a bicycle at certain
speeds, when the bicycle was in 12th gear on flat ground.(Chapter 5)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

