% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biscuit}
\alias{biscuit}
\title{Biscuits}
\format{A data frame with 72 rows and 3 variables:
\describe{
 \item{BLOCK}{oven runs}
 \item{TC}{treatment combination - 9 levels; height of uncooked biscuit (0.25, 0.5,
 0.75 in) x kneading time (7, 14, 21 times)}
 \item{Y}{percentage of original height}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
biscuit
}
\description{
An experiment to study how to make fluffy biscuits. (Chapter 10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

