% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bleach}
\alias{bleach}
\title{Bleach}
\format{A data frame with 45 rows and 4 variables:
\describe{
 \item{CONC}{bleach concentration - 3 levels}
 \item{STAIN}{stain type - 3 levels}
 \item{SPEED}{time for stain removal (in seconds)}
 \item{ORDER}{run order}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
bleach
}
\description{
An experiment to study the effect of different bleach concentrations and
the effect of stain type on speed of stain removal from a piece of cloth.
(Chapter 6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

