% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buttermilk.biscuit}
\alias{buttermilk.biscuit}
\title{Buttermilk biscuits}
\format{A data frame with 30 rows and 4 variables:
\describe{
 \item{BLOCK}{block level}
 \item{POSN}{position in oven - 6 levels}
 \item{BRAND}{brand of biscuit - 3 levels}
 \item{CHANGE}{percentage change in biscuit height}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
buttermilk.biscuit
}
\description{
An experiment to find out which brands of refrigerated buttermilk biscuit
give rise to the fluffiest biscuits. (Chapter 17)
}
\keyword{datasets}

