% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candle}
\alias{candle}
\title{Candle}
\format{A data frame with 64 rows and 3 variables:
\describe{
 \item{BLOCK}{experimenter}
 \item{COLOR}{color of candle - 4 levels (red, white, blue, yellow)}
 \item{TIME}{burning times in seconds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
candle
}
\description{
An experiment to determine whether different colored cake candles burn at different
speeds. (Chapter 10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

