% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalytic.reaction}
\alias{catalytic.reaction}
\title{Catalytic reaction}
\format{A data frame with 16 rows and 6 variables:
\describe{
 \item{RUN}{run order}
 \item{BLOCK}{block identifier - 4 levels}
 \item{A}{converter reaction temperature - 2 levels}
 \item{B}{throughput rate through the converter - 2 levels}
 \item{C}{concentration of the active ingredient in the makeup gas - 2 levels}
 \item{Y}{strength of the product solution}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
catalytic.reaction
}
\description{
An experiment carrying out a catalytic gaseous synthesis reaction to remove the product as a
liquid solution. (Chapter 13, Table 13.27)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

