% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cigarette}
\alias{cigarette}
\title{Cigarette}
\format{A data frame with 60 rows and 6 variables:
\describe{
 \item{order}{time order}
 \item{wp}{whole plot}
 \item{tar}{tar  - 2 levels}
 \item{brand}{brand - 2 levels}
 \item{age}{age - 3 levels}
 \item{time}{burning time in seconds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
cigarette
}
\description{
An experiment to determine factors that affect the length of time that a
cigarette will burn. (Chapter 19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

