% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean.wool}
\alias{clean.wool}
\title{Clean wool}
\format{A data frame with 28 rows and 2 variables:
\describe{
 \item{bale}{bale - 7 levels}
 \item{y}{clean content - percentage the weight of the clean wool is of the
 original weight of the raw wool}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
clean.wool
}
\description{
An experiment to estimate the variability in "clean content" of bales of wool
(Chapter 17)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

