% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coil}
\alias{coil}
\title{Coil}
\format{A data frame with 32 rows and 5 variables:
\describe{
 \item{BLK}{block identifier - 4 levels}
 \item{A}{two winding machines, coded 0, 1}
 \item{B}{two wire socks, coded 0, 1}
 \item{C}{two positions on the coil, coded 0, 1}
 \item{Y}{variability of outside diameters of coils of wire}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
coil
}
\description{
An experiment to examine the variability of outside diameters of coils of wire. (Chapter 13,
Table 13.15)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

