% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colorfastness}
\alias{colorfastness}
\title{Colorfastness}
\format{A data frame with 180 rows and 3 variables:
\describe{
 \item{BLOCK}{experimenter}
 \item{COLOR}{number of times of laundering}
 \item{Y}{color change (continuous version of the Gray Scale for Color Change)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
colorfastness
}
\description{
An experiment to investigate the linear and quadratic effects of the number of
washes on the color change of a denim fabric. (Chapter 10)
}
\keyword{datasets}

