% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotton.spinning}
\alias{cotton.spinning}
\title{Cotton spinning}
\format{A data frame with 78 rows and 5 variables:
\describe{
 \item{BLOCK}{a machine with a single operator}
 \item{TC}{treatment combinations - 6 levels}
 \item{FLYER}{type of flyer - 2 levels}
 \item{TWIST}{degree of twist - 4 levels (1.63, 1.69, 1.78, 1.90)}
 \item{BREAK}{number of breaks per 100 pounds of material}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
cotton.spinning
}
\description{
An experiment to investigate the way in which different degrees of twist (measured
in turns per inch) affect the breakage rate of the roving, and to compare the
ordinary flyer with the newly devised special flyer. (Chapter 2)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

