% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dairy.cow}
\alias{dairy.cow}
\title{Dairy Cow}
\format{A data frame with 18 rows and 4 variables:
\describe{
 \item{COW}{cow identifier}
 \item{PERIOD}{six-week period identifier}
 \item{DIET}{type of diet - 3 levels (roughage, limited grain, full grain)}
 \item{YIELD}{milk yield}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
dairy.cow
}
\description{
An experiment that studies the effect of three diets on the milk production
of dairy cows.
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

