% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decontam.beta}
\alias{decontam.beta}
\title{Decontamination (beta particles)}
\format{A data frame with 32 rows and 6 variables:
\describe{
 \item{BLOCK}{day of the experiment}
 \item{A}{amount of aluminum sulphate added to waste - 2 levels
 (0.4 and 2.5 g per liter)}
 \item{B}{amount of barium chloride added to waste - 2 levels
 (0.4 and 2,5 g per liter)}
 \item{C}{amount of carbon added to waste - 2 levels
 (0.08 and 0.4 g per liter)}
 \item{D}{final pH of waste - 2 levels (6 and 10)}
 \item{COUNT}{counts per minute per milliliter of beta particles}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
decontam.beta
}
\description{
An experiment to explore the effect of four factors on the efficiency of a
decontamination process for the removal of radioactive isotopes from liquid waste.
(Chapter 13)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

