% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dessert}
\alias{dessert}
\title{Dessert}
\format{A data frame with 18 rows and 6 variables
\describe{
 \item{POSITION}{position on table during melting}
 \item{TRTMT}{treatment combinations - 6 levels}
 \item{PCTMELT}{percentage melting by weight}
 \item{SALT}{amount of salt - 2 levels (1/8 and 1/4 tsp)}
 \item{SUGAR}{amount of sugar - 3 levels (1/4, 1/2, and 3/4 cup)}
 \item{FP}{position in freezer - 6 levels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
dessert
}
\description{
An experiment to determine whether melting rate of a frozen dessert will be
affected by the addition of salt and/or sugar. (Chapter 5)
}
\keyword{datasets}

