% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detergent}
\alias{detergent}
\title{Detergent}
\format{A data frame with 36 rows and 5 variables:
\describe{
 \item{BLOCK}{blocking factor - 12 levels}
 \item{TRTMT}{treatment combination - 9 levels}
 \item{Y}{number of plates washed in sink before detergent foam disappeared}
 \item{BASE}{base detergent type - 3 levels}
 \item{ADDITIVE}{parts of additive for base detergent - 4 levels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
detergent
}
\description{
An experiment to compare dishwashing detergent formulations. (Chapter 11, Table 11.8)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

