% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drill.advance}
\alias{drill.advance}
\title{Drill advance}
\format{A data frame with 16 rows and 6 variables:
\describe{
 \item{A}{load on the drill}
 \item{B}{flow rate through the drill}
 \item{C}{speed of rotation}
 \item{D}{type of mud used in drilling}
 \item{ADV}{measure of drill advance}
 \item{LN_ADV}{log to base 10 of drill advance}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
drill.advance
}
\description{
A single replicate 2 x 2 x 2 experiment to study the effects of four treatment
factors on the rate of advance of a small stone drill. (Chapter 7)
}
\examples{
apply(drill.advance[,1:4],2,table) # each factor observed at 2 levels
# set A - D as factor class
for(i in 1:4){
  drill.advance[[i]] <- factor(drill.advance[[i]])
  }
aov1 <- aov(LN_ADV ~ A + B + C + D, data=drill.advance)
summary(aov1)
coef(aov1)
op <- par(mfrow=c(2,2))
for(i in LETTERS[1:4]){
  plot(TukeyHSD(aov1, i))
}
par(op)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

