% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug}
\alias{drug}
\title{Drug}
\format{A data frame with 64 rows and 7 variables
\describe{
 \item{BLOCK}{day of experiment}
 \item{WP}{whole plot identifier}
 \item{B}{dose level of drug - 2 levels}
 \item{C}{time after injection that measurement was taken - 2 levels}
 \item{A}{drug type - 8 levels}
 \item{Y}{amount of fluid (in milliliters) measured in pleural cavity}
 \item{LNY}{log of Y}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
drug
}
\description{
An experiment to evaluate eight drugs for the treatment of arthritis. The
experimental units were rats. (Chapter 19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

