% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dye}
\alias{dye}
\title{Dye}
\format{A data frame with 27 rows and 5 variables:
\describe{
 \item{BLK}{block identifier}
 \item{A}{concentration of base material in the free water in the reaction mixture
  - 3 levels}
 \item{B}{volume of free water in the reaction mixture - 3 levels}
 \item{C}{concentration of two inorganic materials in the free water in the
 reaction mixture - 3 levels}
 \item{Y}{volumes of dyestuff resulting from the chemical reactions}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
dye
}
\description{
An experiment that investigates three reactants in the manufacture of a cotton
dyestuff.
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

