% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erythrocyte}
\alias{erythrocyte}
\title{Erythrocyte}
\format{A data frame with 80 rows and 4 variables
\describe{
 \item{SULF}{levels of sulfamerazine (0, 5, 10, 15)}
 \item{TROUGH}{trough identifier}
 \item{FISH}{fish indentifier}
 \item{COUNT}{number of erythrocites per cubic millimeter of blood}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
erythrocyte
}
\description{
An experiment that investigates the effect of sulfamerazine on the number
of erythrocites per cubic millimeter of blood in brown trout. (Chapter 18)
}
\keyword{datasets}

