% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evaporation}
\alias{evaporation}
\title{Evaporation}
\format{A data frame with 16 rows and 4 variables
\describe{
 \item{A}{placement of water - cup (1) or plate (2)}
 \item{B}{amount of salt added to water - 1 tsp (1) or 3 tsp (3)}
 \item{C}{location in closet - windowsill (1) or floor (2)}
 \item{Y}{amount of water evaporated (100 ml less the amount remaining)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
evaporation
}
\description{
An experiment on the evaporation rate of water under different conditions.
(Chapter 7, Table 7.19)
}
\examples{
aggregate(Y ~ A + B + C, data=evaporation, mean)
summary(aov(Y ~ A + B + C, data=evaporation))
with(evaporation, interaction.plot(A,B,Y))
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

