% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.paper}
\alias{exam.paper}
\title{Exam paper}
\format{A data frame with 130 rows and 4 variables
\describe{
 \item{BLOCK}{block indentifier, accoding to the student's TA}
 \item{A}{color of exam - 2 levels (yellow or green)}
 \item{B}{version of exam - 2 levels}
 \item{Y}{exam score}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
exam.paper
}
\description{
An experiment to see whether student average exam scores differ according to
exam version and color paper. (Chapter 10)
}
\examples{
aggregate(Y ~ A*B, data=exam.paper, mean)
boxplot(Y ~ A*B, data=exam.paper)
summary(lm(Y ~ A*B, data=exam.paper))
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

