% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercise.bicycle}
\alias{exercise.bicycle}
\title{Exercise bicycle}
\format{A data frame with 24 rows and 7 variables:
\describe{
 \item{DAY}{day of experiment}
 \item{SUBJECT}{experimenter identifier}
 \item{DURAT}{time duration of exercice - 2 levels (1 and 3 mins)}
 \item{SPEED}{exercise speed - 2 levels (40 and 60 rpm)}
 \item{PEDAL}{pedal type - 2 levels (foot pedal and hand bars)}
 \item{PULSE}{time in seconds for 50 heart beats}
 \item{TC}{treatment combinations - 8 levels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
exercise.bicycle
}
\description{
An experiment to investigate effects of exercise on pulse rate. (Chapter 12)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

