% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exercise}
\alias{exercise}
\title{Exercise}
\format{A data frame with 36 rows and 7 variables.
\describe{
 \item{BLOCK}{amount of exercise - 3 levels (infrequent, medium, frequent)}
 \item{A}{program setting on bicycle - 2 levels (manual or hill)}
 \item{B}{intensity setting - 2 levels (setting 3 or 5)}
 \item{Y}{difference between post-exercise pulse and pre-exercise pulse}
 \item{AGE}{age of subject}
 \item{SEX}{gender of subject}
 \item{ORDER}{time order}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
exercise
}
\description{
An experiment to examine how pulse rate changes due to different exercise
intensities and programs on a stationary bicycle. (Chapter 10, Table 10.30)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

