% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field}
\alias{field}
\title{Field}
\format{A data frame with 16 rows and 7 variables:
\describe{
 \item{BLK}{block of land}
 \item{A}{amount of dung spread per acre - 2 levels (0 or 10 tons)}
 \item{B}{amount of nitrochalk - 2 levels (0 or 45 lb)}
 \item{C}{amount of superphosphate - 2 levels (0 and 67 lb)}
 \item{D}{amount of muriate of potash - 2 levels (0 and 112 lb)}
 \item{Y}{yield of beans}
 \item{TC}{treatment combination - 2^4 = 16 levels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
field
}
\description{
Results of a field experiment on the yield of beans using various types of
fertilizer. (Chapter 13)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

