% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{film.viscosity}
\alias{film.viscosity}
\title{Film viscosity}
\format{A data frame with 18 rows and 8 variables:
\describe{
 \item{POINT}{design point}
 \item{zP}{coded factor levels of protein concentration}
 \item{xP}{protein concentration}
 \item{zH}{coded factor levels of pH}
 \item{xH}{pH}
 \item{zT}{coded factor levels for temperture (Celsius)}
 \item{xT}{temperture (Celsius)}
 \item{y}{apparent viscosity (mPa)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
film.viscosity
}
\description{
An experiment involving viscosity of film-forming solution. (Chapter 16)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

