% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishing.line}
\alias{fishing.line}
\title{Fishing line}
\format{A data frame with 16 rows and 4 variables:
\describe{
 \item{REEL}{the whole plots}
 \item{BRAND}{brand of fishing line - 2 levels}
 \item{WEIGHT}{stress levels - 2 levels}
 \item{Y}{weight of water that broke line (lb)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
fishing.line
}
\description{
An experiment to compare the strength of two brands of fishing line exposed to two
different levels of stress. (Chapter 19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

