% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flour.expt1}
\alias{flour.expt1}
\title{Flour experiment number 1}
\format{A data frame with 64 rows and 8 variables:
\describe{
 \item{A}{ingredient A - 2 levels}
 \item{B}{ingredient B - 2 levels}
 \item{C}{ingredient C - 2 levels}
 \item{D}{ingredient D - 2 levels}
 \item{E}{ingredient E - 2 levels}
 \item{F}{ingredient F - 2 levels}
 \item{DAY}{noise combinations - 4 levels}
 \item{Y}{yield of flour}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
flour.expt1
}
\description{
An experiment to examine the effect of 6 ingredients on average yield of flour.
(Chapter 15, Table 15.51)
}
\keyword{datasets}

