% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flour.expt3}
\alias{flour.expt3}
\title{Flour experiment number 3}
\format{A data frame with 32 rows and 10 variables:
\describe{
 \item{BLK}{day 1 or 2}
 \item{TC}{treatment combinations - 32 levels}
 \item{Y}{flour yield}
 \item{A1}{pseudofactor 1 of 4-level flour formulation factor}
 \item{A2}{pseudofactor 2 of 4-level flour formulation factor}
 \item{A1A2}{pseudofactor interaction}
 \item{N}{amount of yeast - 2 levels (low or high)}
 \item{S}{proof time - 2 levels (short or long)}
 \item{Q}{degree of mixing and moulding - 2 levels (under or over)}
 \item{T}{dough time delay - 2 levels (short or long)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
flour.expt3
}
\description{
An experiment to find flour that is least variable under different levels of noise
variables. (Chapter 15, Table 15.17)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

