% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flour.production}
\alias{flour.production}
\title{Flour production}
\format{A data frame with 16 rows and 9 variables:
\describe{
 \item{zB}{coded design factor B}
 \item{zC}{coded design factor C}
 \item{zD}{coded design factor D}
 \item{y1}{mean volume of 3 loaves on first half-day}
 \item{y2}{mean volume of 3 loaves on second half-day}
 \item{y3}{mean volume of 3 loaves on third half-day}
 \item{y4}{mean volume of 3 loaves on fourth half-day}
 \item{mean}{mean of y1-y4}
 \item{logs}{log base 10 of \code{mean}}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
flour.production
}
\description{
An experiment to find effect of ingredients on bread loaf volume. (Chapter 16,
Table 16.12)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

