% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fractionation}
\alias{fractionation}
\title{Fractionation}
\format{A data frame with 25 rows and 7 variables:
\describe{
 \item{Run}{run order}
 \item{z1}{extraction time - 2 levels}
 \item{z2}{solvent volume - 2 levels}
 \item{z3}{ethanol concentration - 2 levels}
 \item{z4}{temperature - 2 levels}
 \item{Yield}{purified lecithin yield}
 \item{PCE}{phosphatidylcholine enrichment}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
fractionation
}
\description{
An experiment that studies the effects of extraction time, solvent volume,
ethanol concentration, and temperature on the yield and phosphatidylcholine
enrichment (PCE) of deoiled rapeseed lecithin when fractionated with
ethanol. (Chapter 16, Table 16.25)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

