% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galling}
\alias{galling}
\title{Galling}
\format{A data frame with 8 rows and 4 variables:
\describe{
 \item{A}{speed - 2 levels (1.5 and 5.0 rpm)}
 \item{B}{surface roughness - 2 levels (1.9 and 3.8 microns)}
 \item{C}{axial load - 2 levels (413.4 and 689 MPa)}
 \item{Y}{"galling", or failure phenomenon of severe adhesive wear, on scale of
 0 -10}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
galling
}
\description{
An experiment to study the effects of speed, surface roughness, and axial load
on the "galling" of a metal collar. (Chapter 7, Table 7.23)
}
\keyword{datasets}

