% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasoline.pilot}
\alias{gasoline.pilot}
\title{Gasoline pilot}
\format{A data frame with 9 rows and 4 variables:
\describe{
 \item{CAR}{car/driver block - 3 levels}
 \item{OCTANE}{octane - 3 levels (87, 89, 93)}
 \item{MPG}{miles per gallon}
 \item{ORDER}{order of observation within block}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
gasoline.pilot
}
\description{
An experiment to determine whether the use of a higher-octane gasoline
increases the miles per gallon achieved by an automobile. (Chapter 10,
Table 10.19)
}
\keyword{datasets}

