% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{golf.ball}
\alias{golf.ball}
\title{Golf Ball}
\format{A data frame with 36 rows and 3 variables:
\describe{
 \item{golfer}{person hitting ball - 2 levels}
 \item{brand}{brand of golf ball - 3 levels}
 \item{y}{distance traveled in yards}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
golf.ball
}
\description{
Effect of brand on distance traveled by golf balls. (Chapter 17, Table 17.19)
}
\keyword{datasets}

