% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{handwheel}
\alias{handwheel}
\title{Handwheel}
\format{A data frame with 64 rows and 9 variables:
\describe{
 \item{BLOCK}{block - 4 levels}
 \item{A}{handwheel diameter - 2 levels (5.5 or 10 in)}
 \item{B}{dial diameter - 2 levels (4 or 8 in)}
 \item{C}{thickness of the dial line - 2 levels (0.008 or 0.064 in)}
 \item{D}{friction of the spindle - 2 levels (7.5 or 45 lb.-in)}
 \item{E}{level of operator's elbow relative to height of handwheel - 2 levels
 (level or above)}
 \item{F}{previous experience of operator - 2 levels (practiced or nonpracticed)}
 \item{G}{knowledge of accuracy of previous setting - 2 levels (feedback or no
 feedback)}
 \item{LOGS}{response variable: log of sample variance of 25 repeated observations
 for a particular treatment combination}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
handwheel
}
\description{
An experiment for finding the design of a machine tool handwheel that
would maximize the accuracy on the part of the operator in the setting
of the machine tool handwheel. (Chapter 15, Table 15.48)
}
\keyword{datasets}

