% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heartlung.pump}
\alias{heartlung.pump}
\title{Heart-lung pump}
\format{A data frame with 20 rows and 3 variables:
\describe{
 \item{ORDER}{run order}
 \item{RPM}{number of rpms - 5 levels (50, 75, 100, 125, 150)}
 \item{Y}{fluid flow (liters/minute)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
heartlung.pump
}
\description{
An experiment to determine the effect of the number of revolutions per minute (rpm)
of the rotary pump head of an Olson heart-lung pump on the fluid flow rate.
(Chapter 3, Table 3.2)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

