% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icecream}
\alias{icecream}
\title{Ice cream}
\format{A data frame with 33 rows and 3 variables:
\describe{
 \item{FLAVOR}{flavor of ice cream - 3 levels}
 \item{TIME}{melting time in seconds}
 \ITEM{ORDER}{order of observation}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
icecream
}
\description{
An experiment to determine if different flavors of ice cream melt at different
speeds. (Chapter 17, Table 17.2)
}
\keyword{datasets}

