% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inclinometer.mixed}
\alias{inclinometer.mixed}
\title{Inclinometer mixed}
\format{A dataframe with 216 rows and 15 variables. (See section 15.7.1 for
explanation of variables.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
inclinometer.mixed
}
\description{
An experiment that seeks to improve the performance of an inclinometer,
an instrument that records the angle of tilt of an object such as a crane jib.
(Chapter 15, Table 15.31)
}
\keyword{datasets}

