% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inclinometer.product}
\alias{inclinometer.product}
\title{Inclinometer product}
\format{A dataframe with 27 rows and 17 variables. (See example 15.8.1 for
explanation of variables.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
inclinometer.product
}
\description{
An experiment that seeks to improve the performance of an inclinometer,
an instrument that records the angle of tilt of an object such as a crane jib.
(Chapter 15, Table 15.41)
}
\keyword{datasets}

