% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{injection.molding}
\alias{injection.molding}
\title{Injecton molding}
\format{A data frame with 40 rows and 8 variables:
\describe{
 \item{A}{injection velocity - 2 levels}
 \item{B}{cooling time - 2 levels}
 \item{C}{barrel zone temperature - 2 levels}
 \item{D}{mold temperature - 2 levels}
 \item{E}{hold pressure - 2 levels}
 \item{F}{back pressure - 2 levels}
 \item{LNTH}{length - (deviation from 14.5) * 10^4}
 \item{WDTH}{width - (deviation from 9.35) * 10^4}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
injection.molding
}
\description{
An experiment on the effect of six factors on the shrinkage of a part
produced by injection molding. (Chapter 15, Table 15.52)
}
\keyword{datasets}

