% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ink}
\alias{ink}
\title{Ink}
\format{A data frame with 12 rows and 4 variables:
\describe{
 \item{CLOTH}{cloth type - 3 levels}
 \item{DETGNT}{detergent type - 2 levels}
 \item{ORDER}{order of data collection}
 \item{Y}{stain score, scale of 1 (black) - 19 (white)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
ink
}
\description{
An experiment to determine which type of cloth is most difficult to clean and
whether a detergent plus stain remover was better than a detergent without
stain remover. (Chapter 6, Table 6.22)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

