% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insole.cushion}
\alias{insole.cushion}
\title{Insole cushion}
\format{A data frame with 40 rows and 6 variables:
\describe{
 \item{ORDER}{order of collection}
 \item{BLOCK}{right leg or left leg}
 \item{TRTMT}{treatment combination - 4 levels}
 \item{CUSHN}{cushion type - 2 levels (insole or heel)}
 \item{BRAND}{brand type - 2 levels (brand name or store name)}
 \item{FORCE}{response in Newtons}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
insole.cushion
}
\description{
An experiment to compare the effect on the force with which the foot hits the
ground of a regular insole cushion and a heel cushion, available both as a
brand name and a store name. (Chapter 10, Table 10.28)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

