% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isomer}
\alias{isomer}
\title{Isomer}
\format{A data frame with 30 rows and 3 variables:
\describe{
 \item{block}{block - 10 levels}
 \item{pressure}{pressure - 5 levels (250, 325, 400, 475, and 550 psi)}
 \item{y}{percent conversion}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
isomer
}
\description{
An experiment to obtain information about the effect of pressure on percent
conversion of methyl glucoside to monovinyl isomers. (Chapter 11, Table 11.22)
}
\keyword{datasets}

