% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laser.printer}
\alias{laser.printer}
\title{Laser printer}
\format{A dataframe with 16 rows and 16 variables. (See example pages 534 - 535
for explanation of variables.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
laser.printer
}
\description{
An experiment to determine which options speed up or slow down the printing
of a reasonably detailed picture. (Chapter 15, Table 15.50)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

