% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length.perception}
\alias{length.perception}
\title{Length perception}
\format{A data frame with 84 rows and 5 variables:
\describe{
 \item{SUBJECT}{subject identifier}
 \item{SHAPE}{shape type - 3 levels (square, circle, triangle)}
 \item{AREA}{area type - 2 levels (16 and 9 in^2)}
 \item{TC}{treatment combinations - 6 levels}
 \item{Y}{deviation from target length of 5 cm}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
length.perception
}
\description{
An experiment to explore subjects' ability to reproduce accurately a straight
line of given length (5 cm). (Chapter 10, Table 10.23)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

