% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{light.bulb}
\alias{light.bulb}
\title{Light bulb}
\format{A data frame with 48 rows and 6 variables:
\describe{
 \item{WATTG}{wattage - 2 levels (60 or 100 watt)}
 \item{CAPCTY}{capacity - 2 levels (50 or 100 percent)}
 \item{BRAND}{brand type - 3 levels}
 \item{TRTMT}{treatment combination - 6 levels (BRAND X CAPCTY)}
 \item{RESIST}{observed resistance}
 \item{ORDER}{order of observation}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
light.bulb
}
\description{
An experiment to compare the light intensities of different brands of light bulbs.
(Chapter 10, Table 10.9)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

