% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{load.carrying}
\alias{load.carrying}
\title{Load carrying}
\format{A data frame with 36 rows and 6 variables:
\describe{
 \item{SUBJECT}{subject identifier}
 \item{ORDER}{order of observation within block}
 \item{FLEX}{elbow flexion - 3 levels (0, 45 and 90 degrees)}
 \item{ROTN}{shoulder rotation - 3 levels (0, 45, and 90 degrees)}
 \item{Y}{maximum torque generated at the elbow joint measured in Newtons}
 \item{TC}{treatment combinations - 9 levels (FLEX x ROTN)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
load.carrying
}
\description{
An experiment to investigate the load-carrying capability of the human arm.
(Chapter 10, Table 10.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

