% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ltContrast.R
\name{ltContrast}
\alias{ltContrast}
\title{Calculate linear trend contrast coefficients.}
\usage{
ltContrast(f)
}
\arguments{
\item{f}{A factor.}
}
\description{
Calculates linear trend contrast coefficients according to 4.2.4.
}
\examples{
# example 4.23
is.factor(heartlung.pump$RPM) # TRUE
ltc <- ltContrast(heartlung.pump$RPM)
afr <- with(heartlung.pump, tapply(Y,RPM, mean))
# least squares estimate of linear contrast:
crossprod(afr,ltc)
# get msE:
lmHL <- lm(Y ~ RPM, heartlung.pump)
msE <- summary(lmHL)$sigma^2
# estimated standard error of the linear trend estimator:
sqrt(msE * sum(ltc^2/table(heartlung.pump$RPM)))
}

