% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{machine.head}
\alias{machine.head}
\title{Machine head}
\format{A data frame with 80 rows and 3 variables:
\describe{
 \item{machine}{machine identifier}
 \item{head}{head identifier (nested within machine)}
 \item{y}{strain reading}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
machine.head
}
\description{
An experiment to study the differences in the strain readings of four different
heads on each of five different machines. (Chapter 18, Table 18.3)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

