% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mangold}
\alias{mangold}
\title{Mangold}
\format{A data frame with 32 rows and 7 variables:
\describe{
 \item{BLOCK}{experimental area}
 \item{A}{sulphate of ammonia - 2 levels (0 and 0.6 cwt per acre)}
 \item{B}{superphosphate - 2 levels (0 and 0.5 cwt per acre)}
 \item{C}{muriate of potash - 2 levels (0 and 1.0 cwt per acre)}
 \item{D}{agricultural salt - 2 levels (0 and 5 cwt per acre)}
 \item{E}{dung - 2 levels (0 or 10 tons per acre)}
 \item{Y}{yield of mangold roots in pounds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
mangold
}
\description{
An experiment to investigate the effects of five different fertilizers on the growth of mangold roots.
(Chapter 13, Table 13.11)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

