% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{margarine}
\alias{margarine}
\title{Margarine}
\format{A data frame with 40 rows and 2 variables:
\describe{
 \item{BRAND}{brand of margarine - 4 levels (1 - 3 are margarine brands, 4 is butter)}
 \item{TIME}{time in seconds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
margarine
}
\description{
Melting times in seconds for three different brands of margarine and one brand of butter.
(Chapter 5, Table 5.14)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

