% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memory}
\alias{memory}
\title{Memory}
\format{A data frame with 27 rows and 3 variables:
\describe{
 \item{WDLIST}{word type - 3 levels (fruit, nouns, mixed)}
 \item{DISTRACT}{type of distraction - 3 levels (none, constant, changing)}
 \item{Y}{number of words remembered}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
memory
}
\description{
An experiment to examine the effects of external distractions on short-term memory and to
examine whether some types of words were easier to memorize than others.
(Chapter 6, Table 6.21)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

