% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mung.bean}
\alias{mung.bean}
\title{Mung bean}
\format{A data frame 48 rows and 5 variables:
\describe{
 \item{ORDER}{order of observation}
 \item{WATER}{amount of water - 3 levels (1, 2, or 3 tsp per dat)}
 \item{MAT}{growing medium - 2 levels (tissue or paper towel)}
 \item{LENGTH}{shoot length in mm}
 \item{TRTMT}{treatment combination (WATER x MAT)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
mung.bean
}
\description{
An experiment to compare watering schedules and growing mediums for mung bean seeds.
(Chapter 5, Table 5.7)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

