% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nail.varnish}
\alias{nail.varnish}
\title{Nail varnish}
\format{A data frame with 30 rows and 4 variables:
\describe{
 \item{ORDER}{order of obs}
 \item{SOLUTN}{brand of solvent - 2 levels}
 \item{VARNISH}{brand of varnish - 3 levels}
 \item{TIME}{time in minutes until varnish dissolved}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
nail.varnish
}
\description{
An experiment on the efficacy of nail varnish solvent in removing nail varnish from cloth.
(Chapter 6, Table 6.6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

