% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oats}
\alias{oats}
\title{Oats}
\format{A data frame with 72 rows and 5 variables:
\describe{
 \item{BLOCK}{experimental area - 6 levels}
 \item{WP}{whole plot - 18 levels}
 \item{A}{variety of oat - 3 levels}
 \item{B}{amount of manure - 4 levels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
oats
}
\description{
An experiment on the yield of three varities of oats and four different levels of manure.
(Chapter 19, Table 19.3)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

