% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{operator}
\alias{operator}
\title{Operator}
\format{A data frame with 80 rows and 3 variables:
\describe{
 \item{SAMPLE}{steel sample - 10 levels}
 \item{OPERATOR}{operator - 4 levels}
 \item{CONC}{manganese concentration (percent)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
operator
}
\description{
An experiment to determine how much variation in measured manganese concentration in steel was due to
operator variations. (Chapter 18, Table 18.11)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

