% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pah.recovery}
\alias{pah.recovery}
\title{PAH recovery}
\format{A data frame with 30 rows and 8 variables. (See section 16.7.2 for explanation of variables)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
pah.recovery
}
\description{
An experiment to study the effects of four factors on the total recovery of polycyclic aromatic
hydrocarbons (PAHs) when extracted from soil. (Chapter 16, Table 16.19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

