% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paint.1}
\alias{paint.1}
\title{Paint 1}
\format{A data frame with 8 rows and 12 variables. (See Ch 15, exercise 7(c) for
explanation of variables)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
paint.1
}
\description{
An experiment on the thickness of a paint coating. (Chapter 15, Table 15.49)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{paint.2}}, \code{\link{paint.followup}}
}
\keyword{datasets}

