% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paint.2}
\alias{paint.2}
\title{Paint 2}
\format{A data frame with 32 rows and 7 variables:
\describe{
 \item{A}{belt speed - 2 levels}
 \item{B}{tube width - 2 levels}
 \item{C}{pump pressure - 2 levels}
 \item{D}{paint viscosity - 2 levels}
 \item{E}{tube height - 2 levels}
 \item{F}{heating temperature - 2 levels}
 \item{y}{paint thickness in mm}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
paint.2
}
\description{
An experiment on the thickness of a paint coating. (Chapter 16, Table 16.1)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{paint.1}}, \code{\link{paint.followup}}
}
\keyword{datasets}

