% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paint.followup}
\alias{paint.followup}
\title{Paint followup}
\format{A data frame with 16 rows and 5 variables:
\describe{
 \item{zA}{belt speed - 2 levels}
 \item{zB}{tube width - 2 levels}
 \item{zC}{pump pressure - 2 levels}
 \item{zD}{paint viscosity - 2 levels}
 \item{y}{paint thickness in mm}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
paint.followup
}
\description{
An experiment on the thickness of a paint coating. (Chapter 16, Table 16.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{paint.1}}, \code{\link{paint.2}}
}
\keyword{datasets}

