% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paper.towel.absorbancy}
\alias{paper.towel.absorbancy}
\title{Paper towel absorbancy}
\format{A data frame with 12 rows and 8 variables:
\describe{
 \item{RUN}{order of obs}
 \item{TRTMT}{treatment combination}
 \item{AB}{treatment combination by levels of A and B}
 \item{DROPS}{number of water drops absorbed}
 \item{TIME}{time to collect observation in seconds}
 \item{AREA}{square inches of towel}
 \item{RATE}{rate at which water droplets fell to towel (DROPS/TIME)}
 \item{ABSORB}{number of water drops absorbed per square inch of towel (DROPS/AREA)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
paper.towel.absorbancy
}
\description{
An experiment to compare the effects of two treatment factors, brand and printing, on the absorbancy
of paper towels. (Chapter 9, Table 9.8)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

