% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{papertowel.strength}
\alias{papertowel.strength}
\title{Paper towel strength}
\format{A data frame with 24 rows and 6 variables:
\describe{
 \item{ABC}{treatment combinations}
 \item{A}{amount of liquid - 2 levels (5 or 10 drops)}
 \item{B}{brand of towel - 2 levels}
 \item{C}{type of liquid - 2 levels (beer or water)}
 \item{STRENGTH}{"strength" of paper towel}
 \item{ORDER}{run order}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
papertowel.strength
}
\description{
An experiment to compare relative strengths of two brands of paper towels under varying levels of
moisture saturation and liquid type. (Chapter 7, Table 7.20)
}
\examples{
with(papertowel.strength, interaction.plot(factor(A), factor(B), STRENGTH))
with(papertowel.strength, interaction.plot(factor(A), factor(C), STRENGTH))
with(papertowel.strength, interaction.plot(factor(B), factor(C), STRENGTH))
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

