% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peas}
\alias{peas}
\title{Peas}
\format{A data frame with 24 rows and 5 variables:
\describe{
 \item{block}{which block - 6 levels}
 \item{N}{nitogen - 2 levels}
 \item{P}{phosphate - 2 levels}
 \item{K}{potassium - levels}
 \item{Y}{yield of peas (pounds/plot)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
peas
}
\description{
An experiment to evaluate the effects of nitrogen, phosphate and potassium on the growth of peas.
(Chapter 13, Table 13.25)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link[datasets]{npk}} in the \code{datasets} package
}
\keyword{datasets}

