% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedestrian.light}
\alias{pedestrian.light}
\title{Pedestrian light}
\format{A data frame with 32 rows and 2 variables:
\describe{
 \item{PUSHES}{number of pushes - 4 levels}
 \item{TIME}{time in seconds for walk light to appear}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
pedestrian.light
}
\description{
An experiment to determine whether pushing a certain pedestrian light button had an effect on
waiting for the walk light to appear. (Chapter 3, Table 3.12)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

