% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penicillin}
\alias{penicillin}
\title{Penicillin}
\format{A data frame with 32 rows and 7 variables:
\describe{
 \item{BLOCK}{week of experiment - 2 levels}
 \item{A}{corn steep liquor - 2 levels (2 or 3\% strength)}
 \item{B}{lactose - 2 levels (2 or 3\% strength)}
 \item{C}{precursor - 2 levels (0 or 0.05\%)}
 \item{D}{sodium nitrate - 2 levels (0 or 0.3\%)}
 \item{E}{glucose - 2 levels (0 or 0.5\%)}
 \item{YIELD}{yield of penicillin}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
penicillin
}
\description{
An experiment that investigates the effects of various factors on the yield of penicillin in surface
culture experiments. (Chapter 13, Table 13.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

