% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plasma}
\alias{plasma}
\title{Plasma}
\format{A data frame with 18 rows and 5 variables:
\describe{
 \item{BLOCK}{blocking factor - 6 levels}
 \item{DAY}{day of experiment - 2 levels}
 \item{EXPTR}{experimenter - 3 levels}
 \item{TRTMT}{treatment - 6 levels}
 \item{Y}{height of the water column (mm)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
plasma
}
\description{
An experiment to study the effects of different plasma treatments of certain plastic pipet tips on the
capillary action of the pipets. (Chapter 11, Table 11.10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

