% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popcorn.microwave}
\alias{popcorn.microwave}
\title{Popcorn microwave}
\format{A data frame with 36 rows and 4 variables:
\describe{
 \item{brand}{brand of popcorn - 3 levels}
 \item{power}{power of microwave - 2 levels}
 \item{time}{popping time - 3 levels (4, 4.5, or 5 mins)}
 \item{y}{percent of kernels popped}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
popcorn.microwave
}
\description{
An experiment to compare which brands of microwave popcorn yield the highest proportion of popped kernels.
(Chapter 7, Table 7.2)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

