% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popcorn.robust}
\alias{popcorn.robust}
\title{Popcorn robust}
\format{A data frame with 36 rows and 4 variables:
\describe{
 \item{POP}{brand of popcorn - 3 levels}
 \item{OIL}{type of oil - 2 levels (corn oil or name brand popping oil)}
 \item{TIME}{popping time - 3 levels (1.5, 1.75 or 2 mins)}
 \item{PCPOP}{percentage of popped kernels}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
popcorn.robust
}
\description{
An experiment to examine three brands of popcorn and two types of oil in terms of the percentage of edible
kernels obtained after popping 200 kernels of corn for a certain length of time. (Chapter 7, Table 7.25)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

