% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projectile}
\alias{projectile}
\title{Projectile}
\format{A data frame with 16 rows and 7 variables:
\describe{
 \item{day}{day of data collection - 2 levels}
 \item{run}{order of experiment}
 \item{A}{charge weight - 2 levels}
 \item{B}{projectile weight - 2 levels}
 \item{C}{propellant web - 2 levels}
 \item{D}{weapon - 2 levels}
 \item{y}{projectile velocity}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
projectile
}
\description{
An experiment to study the effects on projectile velocity. (Chapter 13, Table 13.22)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

