% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantity.perception}
\alias{quantity.perception}
\title{Quantity perception}
\format{A data frame of 162 rows and 8 variables. (See ch 12, exercise 10 for explanation of variables)
 \describe{
   \item{OBS}{}
   \item{SUBJ}{}
   \item{ORDER}{}
   \item{TRTMT}{}
   \item{NUMBER}{}
   \item{COLOR}{}
   \item{Y}{}
   \item{COUNT}{}
 }}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
quantity.perception
}
\description{
An experiment to investigate the effect of color on the perception of quantity. (Chapter 12, Table 12.15)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

