% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rail.weld}
\alias{rail.weld}
\title{Rail weld}
\format{A data frame with 12 rows and 4 variables:
\describe{
 \item{T}{ambient temperature - 2 levels (0 and 70 F)}
 \item{V}{wind velocity - 2 levels (0 and 20 mph)}
 \item{S}{rail steel bar size - 2 levels (4/11 and 11/11 inches)}
 \item{Y}{tensile strength}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
rail.weld
}
\description{
An experiment to investigate the effect of three factors on the tensile strength of welds.
(Chapter 7, Table 7.14)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

