% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reaction.time}
\alias{reaction.time}
\title{Reaction time}
\format{A data frame with 18 rows and 5 variables:
\describe{
 \item{ORDER}{order of collection}
 \item{TC}{Treatmebt combination - 6 levels}
 \item{CUE}{cue stimulus - 2 levels (auditory or visual)}
 \item{ELAPTM}{elapsed time between cue and stimulus - 3 levels (5, 10 or 15 secs)}
 \item{REACTM}{reaction time in secs}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
reaction.time
}
\description{
An experiment to compare the effects of auditory and visual cues on speed of response of a human subject.
(Chapter 4, Table 4.6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

