% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refinery}
\alias{refinery}
\title{Refinery}
\format{A data frame with 27 rows and 14 variables. (See example 15.3.1 for explanation of variables)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
refinery
}
\description{
An experiment to find a set of operating conditions to optimize the quality of lube oil treated at a
refinery. (Chapter 15, Table 15.13)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

