% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resin.impurity}
\alias{resin.impurity}
\title{Resin impurity}
\format{A data frame with 13 rows and 4 variables:
\describe{
 \item{point}{design point}
 \item{time}{drying time}
 \item{temp}{temperature in celsius}
 \item{y}{content of undesirable compounds (ppm)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
resin.impurity
}
\description{
An experiment to study the effects of drying time and temperature on the content of undesirable
compounds in a resin. (Chapter 16, Table 16.27)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

