% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resin.moisture}
\alias{resin.moisture}
\title{Resin moisture}
\format{A data frame with 17 rows and 8 variables:
\describe{
 \item{POINT}{design point}
 \item{zT}{coded factor levels for \code{xT}}
 \item{xT}{temperature - 3 levels (150, 185, 220 Celsius)}
 \item{zH}{coded factor levels for \code{xH}}
 \item{xH}{relative humidity - 3 levels (0, 50, 100 \%)}
 \item{zP}{coded factor levels for \code{xP}}
 \item{xP}{air pressure - 3 levels (1, 5, 9 torr)}
 \item{y}{resin degradation (ppm)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
resin.moisture
}
\description{
An experiment to determine whether specific drying conditions for a process could yield a
resin that is sufficiently devoid of moisture and low-molecular-weight components. (Chapter
16, Table 16.28)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

