% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respiratory.exchange.ratio}
\alias{respiratory.exchange.ratio}
\title{Respitory exchange ratio}
\format{A data frame with 27 rows and 3 variables:
\describe{
 \item{SUBJECT}{subject identifier}
 \item{PRTCOL}{protocol type - 3 levels}
 \item{RATE}{respitory exchange ratio}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
respiratory.exchange.ratio
}
\description{
An experiment to compare the effects of inpatient and outpatient protocols on the in-laboratory measurement
of respitory exchange ratio. (Chapter 10, Table 10.18)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{resting.metabolic.rate}}
}
\keyword{datasets}

