% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resting.metabolic.rate}
\alias{resting.metabolic.rate}
\title{Resting metabolic rate}
\format{A data frame with 27 rows and 3 variables:
\describe{
 \item{SUBJECT}{subject identifier}
 \item{PRTCOL}{protocol type - 3 levels}
 \item{RATE}{resting metabolic rate}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
resting.metabolic.rate
}
\description{
An experiment to compare the effects of inpatient and outpatient protocols on the in-laboratory measurement
of resting metabolic rate. (Chapter 10, Table 10.4)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{respiratory.exchange.ratio}}
}
\keyword{datasets}

