% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rocket}
\alias{rocket}
\title{Rocket}
\format{A data frame with 32 rows and 5 variables:
\describe{
 \item{A}{vibration - 2 levels}
 \item{B}{temperture cycling - 2 levels}
 \item{C}{altitude cycling - 2 levels}
 \item{D}{firing temperature/altitude - 4 levels}
 \item{Y}{thrust duration in seconds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
rocket
}
\description{
An experiment on the testing of an auxiliary rocket engine. (Chapter 7, Table 7.21)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

