% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saltwater}
\alias{saltwater}
\title{Saltwater}
\format{A data frame with 45 rows and 3 variables:
\describe{
 \item{BLOCK}{experimenter - 3 levels}
 \item{SALT}{amount of salt - 5 levels (0, 8, 16, 24, and 32 grams)}
 \item{TEMP}{boiling point temperature in Celsius}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
saltwater
}
\description{
An experiment to study the effect of salt on the boiling point of water. (Chapter 10, Table 10.21)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

