% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sludge}
\alias{sludge}
\title{Sludge}
\format{A data frame with 8 rows and 6 variables:
\describe{
 \item{A}{reactor biomass concentration - 2 levels (3000 or 6000 mg/l)}
 \item{B}{clarifier biomass concentration - 2 levels (8000 or 12000 mg/l)}
 \item{C}{waste sludge flow rate - 2 levels (78.5 or 940 m^3/d)}
 \item{D}{biological growth rate constant - 2 levels (0.040 and 0.075 d^-1)}
 \item{E}{fraction of food to biomass - 2 levels (0.4 or 0.8 kg/kg)}
 \item{Y}{removal of biochemical oxygen demand}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
sludge
}
\description{
An experiment that investigates the effect of several factors on the removal of biochemical oxygen demand
from wastewater. (Chapter 15, Table 15.8)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

