% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soap}
\alias{soap}
\title{Soap}
\format{A data frame with 12 rows and 5 variables:
\describe{
 \item{SOAP}{type of soap - 3 levels (regular, deodorant, moisturizing)}
 \item{CUBE}{cube of soap}
 \item{PRE_WT}{weight of cube before soaking}
 \item{POST_WT}{weight of cube after soaking}
 \item{WT_LOSS}{weight loss of soap (POST_WT - PRE_WT)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
soap
}
\description{
An experiment to compare the extent to which three types of soap dissolve in water. (Chapter 2, Table 2.7)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

