% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soup}
\alias{soup}
\title{Soup}
\format{A data frame with 16 rows and 6 variables:
\describe{
 \item{A}{number of mixer ports - 2 levels (1 or 3)}
 \item{B}{temperature of mixture jackets - 2 levels ambient temp or presence of cooling water}
 \item{C}{mixing time - 2 levels (60 or  80 seconds)}
 \item{D}{batch weight - 2 levels (1500 or 2000 lbs)}
 \item{E}{delay between mixing and packaging - 2 levels (1 or 7 days)}
 \item{Y}{measure of variation in weight}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
soup
}
\description{
An experiment to investigate the influence of five factors on soup mix weight variation.
(Chapter 15, Table 15.5)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

