% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spaghetti.sauce}
\alias{spaghetti.sauce}
\title{Spaghetti sauce}
\format{A data frame with 18 rows and 5 variables:
\describe{
 \item{ORDER}{time order}
 \item{BRAND}{brand of sauce - 3 levels}
 \item{STIR}{stirred or unstirred}
 \item{WT}{weight of sauce in grams}
 \item{TRTMT}{treatment combination - 6 levels (see Ch 15, exercise 9 for details)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
spaghetti.sauce
}
\description{
An experiment to compare the thickness of three brands of spaghetti sauce. (Chapter 5, Table 5.20)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

