% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectrometer}
\alias{spectrometer}
\title{Spectrometer}
\format{A data frame with 72 rows and 6 variables:
\describe{
 \item{A}{temperature of lab - 3 levels (67, 72, 77)}
 \item{B}{cleanliness of entrance window seal - 2 levels (clean, one-week use)}
 \item{C}{placement of sample - 3 levels (tangential, covering, partial)}
 \item{D}{wear of boron nitride disk - 2 levels (new, one-month old)}
 \item{E}{sharpness of counterelectrode tip - 2 levels (newly sharpened, one-week wear)}
 \item{Y}{manganese measurements}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
spectrometer
}
\description{
An experiment to determine the causes of instability of measurements made by a Baird spectrometer during
production. (Chapter 7, Table 7.22)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

