% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steel.bar}
\alias{steel.bar}
\title{Steel bar}
\format{A data frame with 96 rows and 4 variables:
\describe{
\item{A}{heat treatment - 2 levels}
\item{B}{screw machine - 4 levels}
\item{C}{time of day - 3 levels (8 am, 11 am, 3 pm)}
\item{Y}{(length - 4.38) x 100 inches}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
steel.bar
}
\description{
An experiment that investigated the cause of variability of the length of steel bars in a manufacturing
process. (Chapter 7, Table 7.26)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

