% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{step}
\alias{step}
\title{Step}
\format{A data frame with 30 rows and 5 variables:
\describe{
 \item{BLOCK}{subject}
 \item{HEIGHT}{step height - 2 levels (5.75 or 11.5 inches)}
 \item{FREQ}{step frequency - 3 levels (14, 21, 28 steps per minute)}
 \item{TRTMT}{treatment combinations - 6 levels}
 \item{PULSE}{pulse rate in beats per minute}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
step
}
\description{
An experiment to investigate the effects on heart rate due to the use of a step machine.
(Chapter 11, Table 11.14)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

