% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sugar.beet}
\alias{sugar.beet}
\title{Sugar beet}
\format{A data frame with 27 rows and 5 variables:
\describe{
 \item{BLOCK}{experimental field block}
 \item{N}{nitogen - 3 levels}
 \item{P}{phosphate - 3 levels}
 \item{K}{potassium - 3 levels}
 \item{Y}{yield of sugar beet}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
sugar.beet
}
\description{
An experiment on the yield of sugar beet. (Chapter 14, Table 14.20)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

