% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival}
\alias{survival}
\title{Survival}
\format{A data frame with 48 rows and 3 variables:
\describe{
 \item{POISON}{amount of poison - 3 levels}
 \item{TRTMT}{treatment - 4 levels}
 \item{TIME}{survival time (units of 10 hours)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
survival
}
\description{
An experiment on the survival times of animals to whom a poison and treatment have been administered.
(Chapter 6, Table 6.23)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

