% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systolic.blood.pressure}
\alias{systolic.blood.pressure}
\title{Systolic blood pressure}
\format{A data frame with 10 rows and 3 variables:
\describe{
 \item{JOG}{jogging time}
 \item{SBP}{systolic blood pressure}
 \item{ORDER}{order of collection}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
systolic.blood.pressure
}
\description{
A pilot experiment to investigate the effect of jogging on systolic blood pressure.
(Chapter 8, Table 8.12)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

