% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Temperature}
\format{A data frame with 24 rows and 4 variables:
\describe{
 \item{THMOM}{thermometer type - 3 levels}
 \item{LOC}{location of use - 2 levels (mouth, under the arm)}
 \item{SUBJ}{subject identifier}
 \item{Y}{time in seconds to register body temperature}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
temperature
}
\description{
An experiment to compare the times required for three different digital
thermometers to register body temperature at two different sites. (Chapter 17,
Table 17.10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

