% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{torque.optimization.1}
\alias{torque.optimization.1}
\title{Torque optimization 1}
\format{A data frame with 36 rows and 4 variables:
\describe{
 \item{A}{anvil type - 3 levels (coined, flat, crowned)}
 \item{B}{rivet diameter - 3 levels (7.0. 7.5, 8.0 mm)}
 \item{M}{mahcine identifier}
 \item{Y}{torque (inch-pounds)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
torque.optimization.1
}
\description{
An experiment to stabilize frame torque of a car seat track. (Chapter 7, Table 7.8)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{torque.optimization.2}}
}
\keyword{datasets}

