% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{torque.optimization.2}
\alias{torque.optimization.2}
\title{Torque optimization 2}
\format{A data frame with 9 rows and 4 variables:
\describe{
 \item{AB}{treatment combinations - 9 levels}
 \item{MEAN}{mean torque at level AB}
 \item{VAR}{variance of torque at level AB}
 \item{LNV}{log of variance of torque at level AB}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
torque.optimization.2
}
\description{
An experiment to stabilize frame torque of a car seat track. (Chapter 7, Table 7.10)
}
\examples{
# deriving torque.optimization.2 from torque.optimization.1
mslns <- function(x){
c(round(mean(x),2),
  round(var(x),2),
  round(log(var(x)),3)
 )
}
g <- with(torque.optimization.1, split(Y, list(B,A)))
t2 <- as.data.frame(do.call(rbind, lapply(g, mslns)))
names(t2) <- c("MEAN","VAR","LNV")
t2$AB <- unique(with(torque.optimization.1,
                interaction(A,B,sep="",lex.order = TRUE)))
rownames(t2) <- NULL
t2 <- t2[,c(4,1:3)]
all.equal(t2, torque.optimization.2, tolerance = 1.5e-3)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{torque.optimization.1}}
}
\keyword{datasets}

