% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trout}
\alias{trout}
\title{Trout}
\format{A data frame with 40 rows and 2 variables:
\describe{
 \item{SULFA}{amount of sulfamerazine added to fish food - 4 levels (0, 5, 10, 15 grams)}
 \item{HEMO}{hemoglobin (grams per 100 ml)}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
trout
}
\description{
An experiment to test the effect of sulfamerazine on hemoglobin in brown trout. (Chapter 3,
Table 3.13)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

