% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{video.game}
\alias{video.game}
\title{Video game}
\format{A data frame with 25 rows and 4 variables:
\describe{
 \item{ORDER}{time order}
 \item{DAY}{day video game played}
 \item{TRTMT}{sound mode - 5 levels}
 \item{Y}{game score}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
video.game
}
\description{
An experiment to evaluate in which of five sound modes a person best plays a certain
video game. (Chapter 12, Table 12.13)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

