% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viscosity}
\alias{viscosity}
\title{Viscosity}
\format{A data frame with 80 rows and 5 variables:
\describe{
 \item{Sample}{sample identifier - 2 levels}
 \item{Aliquot}{aliquot identifier - 10 levels}
 \item{Subali}{subaliquots - 2 levels}
 \item{Part}{the part of the subaliquot - 2 levels}
 \item{Visc}{viscosity determination}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
viscosity
}
\description{
An experiment to determine the viscosity of a polymeric material. (Chapter 18, Table 18.10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

