% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitamin.d}
\alias{vitamin.d}
\title{Vitamin D}
\format{A data frame with 71 rows and 3 variables:
\describe{
 \item{BLOCK}{litter of rats - 18 levels}
 \item{TRTMT}{treatments - 6 levels}
 \item{Y}{response}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
vitamin.d
}
\description{
A study concerning the potency of different doses and preparations of vitamin D. (Chapter 11,
Table 11.26)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

