% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voltage}
\alias{voltage}
\title{Voltage}
\format{A data frame with 160 rows and 4 variables:
\describe{
 \item{SETTING}{setting station - 6 levels}
 \item{REGUL}{regulator identifier}
 \item{TESTING}{testing station - 4 levels}
 \item{VOLTG}{voltages}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
voltage
}
\description{
An experiment on reducing the variability of voltage regulators fitted to motor cars.
(Chapter 18, Table 18.6)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

