% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wafer.1}
\alias{wafer.1}
\title{Wafer 1}
\format{A data frame of 16 rows and 17 variables. (See example 15.6.1 for explanation of
variables.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
wafer.1
}
\description{
An experiment to reduce the variability of the thickness of an "epitaxial layer" deposited
onto silicon wafers during the manufacture of integrated circuit devices. Wafer 1 is
is the orthogonal array presented in Chapter 15, Table 15.25.
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{wafer.2}}
}
\keyword{datasets}

