% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wafer.2}
\alias{wafer.2}
\title{Wafer 2}
\format{A data frame of 16 rows and 10 variables. (See Table 15.24 for explanation of
treatment factors and their levels for the experiment.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
wafer.2
}
\description{
An experiment to reduce the variability of the thickness of an "epitaxial layer" deposited
onto silicon wafers during the manufacture of integrated circuit devices. Wafer 2 is the treatment combinations and response variables presented Chapter 15, Table 15.26.
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\seealso{
\code{\link{wafer.1}}
}
\keyword{datasets}

