% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washing.power}
\alias{washing.power}
\title{Washing power}
\format{A data frame with 27 rows and 4 variables:
\describe{
 \item{A}{sodium carbonate - 3 levels (0, 0.05, or 0.1\%)}
 \item{B}{detergent - 3 levels (0.05, 0.1, or 0.2\%)}
 \item{C}{sodium carboxymethyl cellulose - 3 levels (0, 0.025, or 0.05\%)}
 \item{Y}{reflectance}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
washing.power
}
\description{
An experiment that investigates the washing power of a solution as measured by the
reflectance of pieces of cotton cloth after washing. (Chapter 7, Table 7.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

