% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water.boiling}
\alias{water.boiling}
\title{Water boiling}
\format{A data frame with 48 rows and 4 variables:
\describe{
 \item{BURNER}{burner identifier}
 \item{TSPSALT}{amount of salt added to water}
 \item{TIME}{time taken to boil water}
 \item{ORDER}{run order}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
water.boiling
}
\description{
An experiment to examine the amount of time taken to boil a given amount of water on four
different burners of a stove with various amounts of salt added. (Chapter 6, Table 6.24)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

