% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weathering}
\alias{weathering}
\title{Weathering}
\format{A data frame with 36 rows and 5 variables:
\describe{
 \item{E}{exposure conditions - 2 levels (continuous light, alternating light and
 condensation)}
 \item{A}{exposure levels - 3 levels (1200, 2400, 3600 AFU)}
 \item{D}{direction of cut of the fabric - 2 levels (warp, filling)}
 \item{F}{fabic - 3 levels (polyester, acrylic, nylon)}
 \item{Y}{percent change in breaking strength}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
weathering
}
\description{
An experiment to compare the effects of different types of weathering on the breaking
strength of marine fabrics used for sails. (Chapter 7, Table 7.18)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

