% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight.lifting}
\alias{weight.lifting}
\title{Weight lifting}
\format{A data frame with 66 rows and 4 variables:
\describe{
 \item{ORDER}{run order}
 \item{LEGS}{leg position - 2 levels (straight or bent)}
 \item{WEIGHT}{weight - 3 levels (50, 75, 100 lb)}
 \item{BPFS}{pulse beats per 15 seconds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
weight.lifting
}
\description{
An experiment on the effect of lifting different weights with legs either straight or bent
on pulse rate. (Chapter 6, Table 6.19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

