% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weld.strength}
\alias{weld.strength}
\title{Weld strength}
\format{A data frame with 30 rows and 4 variables:
\describe{
 \item{GAGE}{gage bar setting - 3 levels}
 \item{TIME}{time of welding - 5 levels}
 \item{TRTMT}{treatment combinations - 15 levels}
 \item{STRNTH}{strength of weld}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
weld.strength
}
\description{
An experiment to determine effect of gage bar setting and time of welding on the strength
of a weld in a steel bar. (Chapter 6, Table 6.20)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

