% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welding}
\alias{welding}
\title{Welding}
\format{A data frame with 32 rows and 22 variables. (See example 15.2.2 for explanation of
variables.)}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
welding
}
\description{
An experiment that examines the pull strength of welds resulting from a certain welding
process. The experiment concerns 21 factors. (Chapter 15, Table 15.10)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

