% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildflower}
\alias{wildflower}
\title{Wildflower}
\format{A data frame with 40 rows and 3 variables:
\describe{
 \item{TEMP}{temperature levels - 2 levels (spring, summer)}
 \item{STRAT}{storage methods - 2 levels (stratified, unstratified)}
 \item{Y}{number of germinated seeds}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
wildflower
}
\description{
An experiment to determine whether or not the germination rate of an endangered species of
plant is affected by storage temperature or storage method. (Chapter 5, Table 5.19)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

