% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zinc.plating}
\alias{zinc.plating}
\title{Zinc plating}
\format{A data frame with 12 rows and 3 variables:
\describe{
 \item{VENDOR}{vendor indentifier - 3 levels}
 \item{X}{bracket thickness}
 \item{Y}{zinc plating thickness}
}}
\source{
\url{http://www.wright.edu/~dan.voss/bookdata/data.html}
}
\usage{
zinc.plating
}
\description{
An experiment to examine effect of bracket thickness on zinc plating. (Chapter 9, Table 9.7)
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
}
\keyword{datasets}

